pyproject.toml
[project]
name = "hypercode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "langchain>=0.1.16",
    "langgraph>=0.0.37",
    "langchain-openai>=0.1.3",
    "python-dotenv>=1.0.1",
    "langchain-community>=0.0.34",
    "langchain-core>=0.3.66",
    "langgraph-cli[inmem]>=0.3.3",
    "langgraph-prebuilt>=0.2.3",
    "google-generativeai>=0.8.5",
    "langchain-google-genai>=2.0.10",
]
